Fixed: CreateHost/Refugee method

+++ Tier2/src/main/java/via/sep3/group11/tier2/businessLogic/HousingLogic.java
@@ -1,6 +1,7 @@
 package via.sep3.group11.tier2.businessLogic;
 
 import org.springframework.stereotype.Service;
+import via.sep3.group11.tier2.CommunicationInterfaces.AgreementCommunicationInterface;
 import via.sep3.group11.tier2.CommunicationInterfaces.HostCommunicationInterface;
 import via.sep3.group11.tier2.CommunicationInterfaces.HousingCommunicationInterface;
 import via.sep3.group11.tier2.logicInterfaces.HousingInterface;
@@ -9,7 +10,6 @@ import via.sep3.group11.tier2.shared.DTOs.HousingDTO;
 import via.sep3.group11.tier2.shared.DTOs.HousingIdDTO;
 import via.sep3.group11.tier2.shared.DTOs.HousingListDTO;
 import via.sep3.group11.tier2.shared.domain.Address;
-import via.sep3.group11.tier2.shared.domain.Host;
 import via.sep3.group11.tier2.shared.domain.Housing;
 
 import java.util.List;
@@ -20,25 +20,29 @@ public class HousingLogic implements HousingInterface {
 
     private HostCommunicationInterface hostDAO;
     private HousingCommunicationInterface housingDAO;
+    private AgreementCommunicationInterface agreementDAO;
 
-    public HousingLogic(HostCommunicationInterface hostDAO, HousingCommunicationInterface housingDAO) {
+    public HousingLogic(HostCommunicationInterface hostDAO, HousingCommunicationInterface housingDAO,AgreementCommunicationInterface agreementDAO) {
         this.hostDAO = hostDAO;
         this.housingDAO = housingDAO;
+        this.agreementDAO = agreementDAO;
     }
     @Override
     public HousingDTO addHousing(HousingCreationDTO dto) {
-        // Create housing
+
         Housing housingToAdd = new Housing(dto.getCapacity(),
                 new Address(dto.getCountry(), dto.getCity(), dto.getStreetName(),
                         dto.getHouseNumber(), dto.getRoomNumber(), dto.getPostCode()),false);
 
-        // Get Housing from database
-        Housing addedHousing = housingDAO.addHousing(housingToAdd, dto.getHostEmail());
-        // TODO catch exception, if DB wasn't contacted
-        HousingDTO housingDTO = new HousingDTO(addedHousing,"");
-
-
-        return null;
+        try {
+           Housing addedHousing = housingDAO.addHousing(housingToAdd, dto.getHostEmail());
+            return new HousingDTO(addedHousing,"");
+        }
+        catch (Exception e)
+        {
+            System.out.println("Failed connection with data-tier while executing 'addHousing'.");
+            return new HousingDTO(housingToAdd,"Connection to the server was lost.");
+        }
     }
 
     @Override
@@ -55,6 +59,10 @@ public class HousingLogic implements HousingInterface {
         if (housing.isEmpty())
         {
             return new HousingDTO(dummyHousing,"Housing with ID: "+dto.getHousingId()+ " doesn't exist." );
+        }
+        if (!agreementDAO.getAllAgreementsByHousingId(housing.get().getHousingId()).isEmpty())
+        {
+            return new HousingDTO(dummyHousing, "This housing has active contract. First terminate contract.");
         }
         housingDAO.removeHousing(dto.getHousingId());
         return new HousingDTO(housing.get(),"");
